backend/requirements.txt
flask
flask-cors
python-dotenv
requests
PyJWT
bcrypt
openpyxl
cloudinary
omnidimension
waitress
psycopg2-binary
groq
pandas